очивка3/requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
mysql-connector-python==8.0.27
Flask-Cors==3.0.10
